[Refactor]#17 - View와 ViewController 분리
